Base rollup circuit wasm wrapper
