+++ yarn-project/circuits.js/package.json
@@ -36,9 +36,11 @@
   },
   "dependencies": {
     "@aztec/foundation": "workspace:^",
+    "@types/lodash.times": "^4.3.7",
     "cross-fetch": "^3.1.5",
     "detect-node": "^2.1.0",
     "eslint": "^8.35.0",
+    "lodash.times": "^4.3.2",
     "tslib": "^2.4.0"
   },
   "devDependencies": {
