--- yarn-project/circuits.js/src/rollup/base_rollup_wasm.ts
@@ -1,1 +1,0 @@
-export class BaseRollupWasm {}
